chore(app-core): rename src/app-shell/ → src/slots/, src/shell/ → src/desktop-runtime/ (Hard win #5)
Layer 7 audit "Shell seam clarification" — three folders had `shell` in
their name but distinct roles, making the seam ambiguous:

  Before:
    src/app-shell/         Slot registry (1 file)
    src/shell/             Renderer runtime entries (6 files)
    src/components/shell/  In-window React chrome (18 files)

  After:
    src/slots/             ← was src/app-shell/
    src/desktop-runtime/   ← was src/shell/
    src/components/shell/  unchanged (correct name — in-window chrome)

The renames make the role self-documenting: `slots` for the slot
registry pattern (task-coordinator-slots is the first; the audit
notes more will follow), `desktop-runtime` for the Electrobun-spawned
renderer roots (AppWindowRenderer, DetachedShellRoot, three
DesktopRuntime files, index).

Updated subpath consumers:
- packages/app-core/package.json: `./shell` → `./desktop-runtime`,
  `./app-shell/task-coordinator-slots` → `./slots/task-coordinator-slots`
- packages/app-core/src/index.ts + browser.ts: barrel re-exports flipped
- 3 internal relative imports in components/pages/{ChatView,
  TasksPageView}.tsx and shell→desktop-runtime/DetachedShellRoot.tsx

Verified zero external `@elizaos/app-core/shell` or `/app-shell/`
subpath consumers in plugins/, packages/, scripts/ before renaming
(grep across `*.ts`, `*.tsx`, `*.json` excluding node_modules / dist /
build). The standalone `src/app-shell-components.ts` file is unchanged
— it's a separate barrel (file, not folder) that the audit didn't
flag for rename.

Net for this PR: 12 files affected, 11 insertions, 12 deletions.
Pure mechanical rename — typecheck clean across packages/app-core,
biome clean. No behaviour change. The seam is now self-documenting.
