+++ packages/app-core/package.json
@@ -123,7 +123,7 @@
     "./utils/format": "./src/utils/format.ts",
     "./voice": "./src/voice/index.ts",
     "./app-shell-components": "./src/app-shell-components.ts",
-    "./app-shell/task-coordinator-slots": "./src/app-shell/task-coordinator-slots.tsx",
+    "./slots/task-coordinator-slots": "./src/slots/task-coordinator-slots.tsx",
     "./autonomy": "./src/autonomy/index.ts",
     "./chat": "./src/chat/index.ts",
     "./chat/coding-agent-session-state": "./src/chat/coding-agent-session-state.ts",
@@ -152,7 +152,7 @@
     "./styles/xterm.css": "./src/styles/xterm.css",
     "./styles/onboarding-game.css": "./src/styles/onboarding-game.css",
     "./styles/styles.css": "./src/styles/styles.css",
-    "./shell": "./src/shell/index.ts",
+    "./desktop-runtime": "./src/desktop-runtime/index.ts",
     "./styles/brand-gold.css": "./src/styles/brand-gold.css",
     "./character-catalog": "./src/character-catalog.ts",
     "./test-support/test-helpers": "./src/test-support/test-helpers.ts",
