+++ packages/app-core/src/slots/task-coordinator-slots.tsx
